# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Public REST API for agentserver. Generated from\nGo handler annotations via swaggo/swag; see\ndocs/api/README.md for regeneration instructions.",
    "title": "agentserver API",
    "contact": {},
    "version": "0.1.0"
  },
  "paths": {
    "/api/agent/discovery/agents": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Discover agents in the calling agent's workspace (proxy_token auth)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentCardItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent/mailbox/inbox": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Read messages from the calling agent's inbox",
        "parameters": [
          {
            "description": "Max messages to return (default 10)",
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentMailboxMessage"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/mailbox/send": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Send a message to another agent's mailbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentMailboxSendRequest"
              }
            }
          },
          "description": "Message payload",
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentMailboxSendResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "target not in same workspace",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "target agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/register": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Register an agent (obtain sandbox credentials)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegisterRequest"
              }
            }
          },
          "description": "Agent registration info",
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "no permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/tasks": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Create a delegated task (proxy_token auth)",
        "requestBody": {
          "$ref": "#/components/requestBodies/AgentTaskCreateRequest"
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "target agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent/tasks/{id}": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get a task by ID (proxy_token auth)",
        "parameters": [
          {
            "description": "Task ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskDetail"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "requestBodies": {
...
      "AgentTaskCreateRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AgentTaskCreateRequest"
            }
          }
        },
        "description": "Task details",
        "required": true
      }
    },
    "securitySchemes": {
      "CookieAuth": {
        "description": "Session cookie set by POST /api/auth/login. All non-auth endpoints assume this cookie.",
        "type": "apiKey",
        "name": "agentserver-token",
        "in": "cookie"
      }
    },
    "schemas": {
      "AgentCardItem": {
        "type": "object",
        "required": [
          "agent_id",
          "agent_type",
          "card",
          "display_name",
          "status",
          "version"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "agent_type": {
            "type": "string",
            "example": "claudecode"
          },
          "card": {},
          "description": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "available"
          },
          "version": {
            "type": "integer"
          }
        }
      },
...
      "AgentMailboxMessage": {
        "type": "object",
        "required": [
          "created_at",
          "from",
          "id",
          "msg_type",
          "text"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "msg_type": {
            "type": "string",
            "example": "text"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "AgentMailboxSendRequest": {
        "type": "object",
        "required": [
          "text",
          "to"
        ],
        "properties": {
          "msg_type": {
            "type": "string",
            "example": "text"
          },
          "text": {
            "type": "string",
            "example": "Hello from another agent"
          },
          "to": {
            "type": "string",
            "example": "sandbox-uuid-of-target"
          }
        }
      },
      "AgentMailboxSendResponse": {
        "type": "object",
        "required": [
          "message_id",
          "status"
        ],
        "properties": {
          "message_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "sent"
          }
        }
      },
      "AgentRegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Local Agent"
          },
          "type": {
            "description": "optional; defaults to \"opencode\"",
            "type": "string",
            "example": "opencode"
          }
        }
      },
      "AgentRegisterResponse": {
        "type": "object",
        "required": [
          "proxy_token",
          "sandbox_id",
          "short_id",
          "tunnel_token",
          "workspace_id"
        ],
        "properties": {
          "proxy_token": {
            "type": "string"
          },
          "sandbox_id": {
            "type": "string"
          },
          "short_id": {
            "type": "string"
          },
          "tunnel_token": {
            "type": "string"
          },
          "workspace_id": {
            "type": "string"
          }
        }
      },
...
      "AgentTaskCreateRequest": {
        "type": "object",
        "required": [
          "prompt",
          "target_id"
        ],
        "properties": {
          "delegation_chain": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "max_budget_usd": {
            "type": "number",
            "example": 1
          },
          "max_turns": {
            "type": "integer",
            "example": 50
          },
          "prompt": {
            "type": "string",
            "example": "Run the test suite"
          },
          "requester_id": {
            "type": "string"
          },
          "skill": {
            "type": "string",
            "example": "testing"
          },
          "system_context": {
            "type": "string"
          },
          "target_id": {
            "type": "string",
            "example": "sandbox-uuid"
          },
          "timeout_seconds": {
            "type": "integer",
            "example": 300
          }
        }
      },
      "AgentTaskCreateResponse": {
        "type": "object",
        "required": [
          "status",
          "task_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "task_id": {
            "type": "string"
          }
        }
      },
      "AgentTaskDetail": {
        "type": "object",
        "required": [
          "created_at",
          "prompt",
          "status",
          "target_id",
          "task_id",
          "workspace_id"
        ],
        "properties": {
          "completed_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "failure_reason": {
            "type": "string",
            "nullable": true
          },
          "num_turns": {
            "type": "integer"
          },
          "prompt": {
            "type": "string"
          },
          "requester_id": {
            "type": "string"
          },
          "result": {
            "nullable": true
          },
          "session_id": {
            "type": "string",
            "nullable": true
          },
          "skill": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "pending"
          },
          "target_id": {
            "type": "string"
          },
          "task_id": {
            "type": "string"
          },
          "total_cost_usd": {
            "type": "number",
            "nullable": true
          },
          "workspace_id": {
            "type": "string"
          }
        }
      },
...
    }
  }
}